move create_replication_user in Primary class
